scripts for backlink maintenance
This script will add a section for backlinks at the bottom of any note that
is linked to by another note. This is helpful for automating maintenance
of seeing where else in the repo you reference this note.

With the use of the Markdown Links extension, you'll be able to see a
graph of how different notes in your repo are related.