scripts for tag maintenance
It can be nice to tag notes by topic. The tag maintenance scripts automates
creating a way to find all notes for a specific tag. The script creates a
`tags` directory inside the specified directory and creates a file in `tags`
for each unique tag found in your notes. The file has a bulleted list linking
to each note that uses that tag.

Currently, it finds tags via a heading 3 in the note:

```markdown

comma separated, list, of, tags

this content won't get picked up
```